move `get_app_identifier` out
