+++ tests/apps/app/commands/test_new.py
@@ -3,7 +3,11 @@ import shutil
 import tempfile
 from unittest import TestCase
 
-from piccolo.apps.app.commands.new import module_exists, new
+from piccolo.apps.app.commands.new import (
+    get_app_identifier,
+    module_exists,
+    new,
+)
 
 
 class TestModuleExists(TestCase):
@@ -43,3 +47,21 @@ class TestNewApp(TestCase):
                 "A module called sys already exists"
             )
         )
+
+
+class TestGetAppIdentifier(TestCase):
+
+    def test_get_app_identifier(self):
+        """
+        Make sure the the ``root`` argument is handled correctly.
+        """
+        self.assertEqual(
+            get_app_identifier(app_name="music", root="."),
+            "music.piccolo_app",
+        )
+
+        for root in ("apps", "./apps", "./apps/"):
+            self.assertEqual(
+                get_app_identifier(app_name="music", root=root),
+                "apps.music.piccolo_app",
+            )
